Reusable speed optimization mechanism
Last time, I refactored the Stoch function to be able to reuse GetRsi results, then discovered that you were working on ReusableResult, toward the same end.

This is another speed optimization that should work for any(most?/many?) indicators that use quotes occurring on, or before, the result's date (lookbacks only), and may be applicable in many of the places that you intend to use ReusableResult and/or implement chaining. Together, the demand from bots watching many indicators, on many indicators, of many symbols, for many time frames, is drastically reduced. The waithandles for the threads that used to run non-stop, never fast enough, usually timeout, now.

I refactored the parameters for CalcRsi() into a state object, and new CalcRsi, and GetRsi override, and extension methods, were added. Backward-compatibility was maintained.

In a (presumably) typical use-case:
1) Every (1-10 seconds), only the latest quote changes. While iterating, the old CalcRsi function now saves its' state, right after calculating the next-to-last result.
-The state object is discarded by the old public functions, but can be saved by the new ones, and used later to resume the calculation, recalculating only the last result, swapping out the old last quote for newLastQuote.
-The time saved is proportional to the length of the quote list, N-1.
-This is intended to happen only when CalcRsi detects that the last quote and newLastQuote have the same value for Date.
-Implemented, but not tested vigorously

2) A single new quote is added every 1-[interval].
-The saved state now resumes at Quote[Count-3], and is brought current by recalculating Result[Count-2] (same result as before)
-Calculate new last result from Quote[Count-1].
-The time saved is proportional to the length of the quote list, N-2.
-This is intended to happen only when CalcRsi detects that the last quote and newLastQuote have different dates.
-Assertion that the timespan between the two be exactly one [interval] is probably desired/required.
-Currently, I avoid the exception by pre-detecting the condition in the calling code, and calling the full-recalculation method, instead.
-I haven't analyzed the "initialize average gain" branch of the if statement, to determine if I'll need to take Quote[(Count - 1)-LookBackPeriods] out of rsiState.avgGain/Loss, since Quote[Count-LookBackPeriods] refers to the quote after the one used last time.
-Not implemented. An exception is thrown.